feat: accepted a new structur
